TEST COPY COMPETITOR OBJECT
Race.addCompetitor() stores a Member object as instance variable in the Race object.
If we change the original Member object outside the Race object it is also changed inside the Race object!!
We need a deep copy to preserve the Member instance variable in the Race object.
Is here tested and demonstrated in Main